Re-org, use latest cosmwasm-schema features, regen schema / ts
