+++ contracts/pre-propose/cwd-pre-propose-approval-single/examples/schema.rs
@@ -1,12 +1,10 @@
 use cosmwasm_schema::write_api;
-use cosmwasm_std::Empty;
-use cwd_pre_propose_approval_single::ProposeMessage;
-use cwd_pre_propose_base::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use cwd_pre_propose_approval_single::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
 
 fn main() {
     write_api! {
-        instantiate: InstantiateMsg<Empty>,
-        query: QueryMsg<Empty>,
-        execute: ExecuteMsg<ProposeMessage, Empty>,
+        instantiate: InstantiateMsg,
+        query: QueryMsg,
+        execute: ExecuteMsg,
     }
 }
