+++ contracts/pre-propose/cwd-pre-propose-approval-single/src/state.rs
@@ -2,7 +2,7 @@ use cosmwasm_schema::cw_serde;
 use cosmwasm_std::Addr;
 use cw_storage_plus::{Item, Map};
 
-use crate::contract::ProposeMessageInternal;
+use crate::msg::ProposeMessageInternal;
 
 #[cw_serde]
 pub struct PendingProposal {
