+++ contracts/pre-propose/cwd-pre-propose-approver/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod contract;
+pub mod msg;
+pub mod state;
 
 #[cfg(test)]
 mod tests;
-
-pub use contract::{ExecuteMsg, InstantiateMsg, QueryMsg};
 
 // Exporting these means that contracts interacting with this one don't
 // need an explicit dependency on the base contract to read queries.
